# api-specification.yaml
openapi: 3.0.3
info:
  title: 加拿大快递配送区域地图系统 API
  description: 提供区域管理、邮编配置、价格设置等功能的RESTful API
  version: 2.0.0
  contact:
    name: API Support
    email: api-support@canadapostal.com

servers:
  - url: http://localhost:3000/api/v1
    description: 开发环境
  - url: https://api.canadapostal.com/v1
    description: 生产环境

security:
  - BearerAuth: []

paths:
  # 认证相关
  /auth/login:
    post:
      tags: [Authentication]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      user:
                        $ref: '#/components/schemas/User'
                      expiresIn:
                        type: string
                        example: 24h

  /auth/logout:
    post:
      tags: [Authentication]
      summary: 用户登出
      responses:
        '200':
          description: 登出成功

  # 区域管理
  /regions:
    get:
      tags: [Regions]
      summary: 获取所有配送区域
      parameters:
        - name: include_inactive
          in: query
          schema:
            type: boolean
            default: false
          description: 是否包含非活跃区域
      responses:
        '200':
          description: 成功获取区域列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeliveryRegion'

    post:
      tags: [Regions]
      summary: 创建新的配送区域
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegionRequest'
      responses:
        '201':
          description: 区域创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DeliveryRegion'

  /regions/{regionId}:
    get:
      tags: [Regions]
      summary: 获取指定区域详情
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: string
            example: "1"
      responses:
        '200':
          description: 成功获取区域详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DeliveryRegionDetail'

    put:
      tags: [Regions]
      summary: 更新区域信息
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRegionRequest'
      responses:
        '200':
          description: 区域更新成功

    delete:
      tags: [Regions]
      summary: 删除区域
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 区域删除成功

  # 邮编管理
  /regions/{regionId}/postal-codes:
    get:
      tags: [Postal Codes]
      summary: 获取区域的邮编列表
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          schema:
            type: string
          description: 搜索FSA代码
      responses:
        '200':
          description: 成功获取邮编列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/PostalCode'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      tags: [Postal Codes]
      summary: 添加邮编到区域
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fsaCodes]
              properties:
                fsaCodes:
                  type: array
                  items:
                    type: string
                    pattern: '^[A-Z][0-9][A-Z]$'
                  example: ["V6B", "V6C"]
      responses:
        '201':
          description: 邮编添加成功

  /postal-codes/{postalCodeId}:
    put:
      tags: [Postal Codes]
      summary: 更新邮编信息
      parameters:
        - name: postalCodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostalCodeRequest'
      responses:
        '200':
          description: 邮编更新成功

    delete:
      tags: [Postal Codes]
      summary: 删除邮编
      parameters:
        - name: postalCodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 邮编删除成功

  # 重量区间管理
  /regions/{regionId}/weight-ranges:
    get:
      tags: [Weight Ranges]
      summary: 获取区域的重量区间
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功获取重量区间列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WeightRange'

    post:
      tags: [Weight Ranges]
      summary: 创建重量区间
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWeightRangeRequest'
      responses:
        '201':
          description: 重量区间创建成功

  # 价格计算
  /calculate-price:
    post:
      tags: [Pricing]
      summary: 计算配送价格
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [regionId, weight]
              properties:
                regionId:
                  type: string
                  example: "1"
                weight:
                  type: number
                  format: float
                  example: 12.5
      responses:
        '200':
          description: 价格计算成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      price:
                        type: number
                        format: float
                        example: 25.50
                      currency:
                        type: string
                        example: CAD
                      weightRange:
                        $ref: '#/components/schemas/WeightRange'

  # 数据导入导出
  /export/regions:
    get:
      tags: [Data Management]
      summary: 导出区域数据
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, excel]
            default: json
      responses:
        '200':
          description: 数据导出成功

  /import/postal-codes:
    post:
      tags: [Data Management]
      summary: 批量导入邮编
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                regionId:
                  type: string
      responses:
        '200':
          description: 导入成功

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          type: string
          enum: [admin, manager, user]
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    DeliveryRegion:
      type: object
      properties:
        id:
          type: string
          example: "1"
        name:
          type: string
          example: "区域1"
        description:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer
        colorCode:
          type: string
          example: "#3B82F6"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    DeliveryRegionDetail:
      allOf:
        - $ref: '#/components/schemas/DeliveryRegion'
        - type: object
          properties:
            postalCodeCount:
              type: integer
            weightRangeCount:
              type: integer
            activePostalCodes:
              type: integer
            activeWeightRanges:
              type: integer

    PostalCode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        regionId:
          type: string
        fsaCode:
          type: string
          pattern: '^[A-Z][0-9][A-Z]$'
          example: "V6B"
        province:
          type: string
          example: "BC"
        city:
          type: string
          example: "Vancouver"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time

    WeightRange:
      type: object
      properties:
        id:
          type: string
          format: uuid
        regionId:
          type: string
        rangeName:
          type: string
          example: "0-11.000 KGS"
        minWeight:
          type: number
          format: float
        maxWeight:
          type: number
          format: float
        price:
          type: number
          format: float
        isActive:
          type: boolean
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time

    CreateRegionRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        colorCode:
          type: string
        displayOrder:
          type: integer

    UpdateRegionRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        colorCode:
          type: string
        displayOrder:
          type: integer

    UpdatePostalCodeRequest:
      type: object
      properties:
        fsaCode:
          type: string
          pattern: '^[A-Z][0-9][A-Z]$'
        province:
          type: string
        city:
          type: string
        isActive:
          type: boolean

    CreateWeightRangeRequest:
      type: object
      required: [rangeName, minWeight, maxWeight, price]
      properties:
        rangeName:
          type: string
        minWeight:
          type: number
          format: float
        maxWeight:
          type: number
          format: float
        price:
          type: number
          format: float
        displayOrder:
          type: integer

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object